Moved everything related to savegames into persistence namespace
