+++ Persistence/CivilianFactionExternalData.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SKCivilianIndustry
+namespace SKCivilianIndustry.Persistence
 {
     public class CivilianFactionExternalData : IArcenExternalDataPatternImplementation
     {
