+++ Persistence/CivilianPlanetExternalData.cs
@@ -1,17 +1,21 @@
-using Arcen.AIW2.Core;
-using Arcen.Universal;
+﻿using Arcen.Universal;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace SKCivilianIndustry
+namespace SKCivilianIndustry.Persistence
 {
-    public class CivilianWorldExternalData : IArcenExternalDataPatternImplementation
+    public class CivilianPlanetExternalData : IArcenExternalDataPatternImplementation
     {
         // Make sure you use the same class name that you use for whatever data you want saved here.
-        private CivilianWorld Data;
+        private CivilianPlanet Data;
 
         public static int PatternIndex;
 
         // So this is essentially what type of thing we're going to 'attach' our class to.
-        public static string RelatedParentTypeName = "World";
+        public static string RelatedParentTypeName = "Planet";
 
         public void ReceivePatternIndex(int Index)
         {
@@ -23,26 +27,25 @@ namespace SKCivilianIndustry
         }
         public bool GetShouldInitializeOn(string ParentTypeName)
         {
-            // Figure out which object type has this sort of ExternalData (in this case, World)
+            // Figure out which object type has this sort of ExternalData (in this case, Faction)
             return ArcenStrings.Equals(ParentTypeName, RelatedParentTypeName);
         }
 
         public void InitializeData(object ParentObject, object[] Target)
         {
-            this.Data = new CivilianWorld();
+            this.Data = new CivilianPlanet();
             Target[0] = this.Data;
         }
         public void SerializeExternalData(object[] Source, ArcenSerializationBuffer Buffer)
         {
             //For saving to disk, translate this object into the buffer
-            CivilianWorld data = (CivilianWorld)Source[0];
+            CivilianPlanet data = (CivilianPlanet)Source[0];
             data.SerializeTo(Buffer);
         }
         public void DeserializeExternalData(object ParentObject, object[] Target, int ItemsToExpect, ArcenDeserializationBuffer Buffer)
         {
             //reverses SerializeData; gets the date out of the buffer and populates the variables
-            Target[0] = new CivilianWorld(Buffer);
+            Target[0] = new CivilianPlanet(Buffer);
         }
     }
- 
 }
