+++ Persistence/CivilianWorldExternalData.cs
@@ -1,21 +1,17 @@
-﻿using Arcen.Universal;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Arcen.AIW2.Core;
+using Arcen.Universal;
 
-namespace SKCivilianIndustry
+namespace SKCivilianIndustry.Persistence
 {
-    public class CivilianPlanetExternalData : IArcenExternalDataPatternImplementation
+    public class CivilianWorldExternalData : IArcenExternalDataPatternImplementation
     {
         // Make sure you use the same class name that you use for whatever data you want saved here.
-        private CivilianPlanet Data;
+        private CivilianWorld Data;
 
         public static int PatternIndex;
 
         // So this is essentially what type of thing we're going to 'attach' our class to.
-        public static string RelatedParentTypeName = "Planet";
+        public static string RelatedParentTypeName = "World";
 
         public void ReceivePatternIndex(int Index)
         {
@@ -27,25 +23,26 @@ namespace SKCivilianIndustry
         }
         public bool GetShouldInitializeOn(string ParentTypeName)
         {
-            // Figure out which object type has this sort of ExternalData (in this case, Faction)
+            // Figure out which object type has this sort of ExternalData (in this case, World)
             return ArcenStrings.Equals(ParentTypeName, RelatedParentTypeName);
         }
 
         public void InitializeData(object ParentObject, object[] Target)
         {
-            this.Data = new CivilianPlanet();
+            this.Data = new CivilianWorld();
             Target[0] = this.Data;
         }
         public void SerializeExternalData(object[] Source, ArcenSerializationBuffer Buffer)
         {
             //For saving to disk, translate this object into the buffer
-            CivilianPlanet data = (CivilianPlanet)Source[0];
+            CivilianWorld data = (CivilianWorld)Source[0];
             data.SerializeTo(Buffer);
         }
         public void DeserializeExternalData(object ParentObject, object[] Target, int ItemsToExpect, ArcenDeserializationBuffer Buffer)
         {
             //reverses SerializeData; gets the date out of the buffer and populates the variables
-            Target[0] = new CivilianPlanet(Buffer);
+            Target[0] = new CivilianWorld(Buffer);
         }
     }
+ 
 }
