+++ SpecialFaction_SKCivilianIndustry.cs
@@ -1,6 +1,7 @@
 ﻿using Arcen.AIW2.Core;
 using Arcen.AIW2.External;
 using Arcen.Universal;
+using SKCivilianIndustry.Persistence;
 using System;
 using System.Collections.Generic;
 using System.Linq;
